feat: Move comps
