+++ .eslintrc.json
@@ -56,6 +56,7 @@
     "jsx-a11y/no-static-element-interactions": ["off"],
     "jsx-a11y/click-events-have-key-events": "off",
     "react/jsx-curly-brace-presence": "off",
-    "@typescript-eslint/no-var-requires": "off"
+    "@typescript-eslint/no-var-requires": "off",
+    "react/react-in-jsx-scope": "off"
   }
 }
