+++ src/components/basis/Content/index.tsx
@@ -3,8 +3,8 @@ import classNames from 'classnames';
 import { Container } from 'reactstrap';
 import { Switch, Route } from 'react-router-dom';
 import './style.css';
-import Topbar from '../basis/TopBar';
-import { ProductPage } from '../../page/index';
+import Topbar from '../TopBar';
+import { ProductPage } from '../../../page/index';
 
 type ContentProps = {
   sidebarIsOpen: boolean;
