+++ src/components/index.ts
@@ -1,4 +1,4 @@
-export { default as Content } from './Content';
+export { default as Content } from './basis/Content';
 export { default as SideBar } from './basis/SideBar';
 export { default as SubMenu } from './basis/SubMenu';
 export { default as TopBar } from './basis/TopBar';
